fix: move internal tx interface to own file
